refactor: extract essay-detection to standalone module + wire CI
Moves NON_QUESTION_PATTERNS, MIN_ESSAY_LABEL_LENGTH, and isEssayQuestion
out of detector.ts (which has browser side effects at module load) into a
pure sibling module essay-detection.ts. Tests can now import the helpers
directly under node --test --experimental-strip-types, dropping the
esbuild + node:vm + chrome stub scaffolding.

Adds an "npm test" script and wires it into the extension CI job after
build. Bumps the extension job to Node 22 for --experimental-strip-types.
